use component to render links in sidebar
