+++ src/components/v2/Layout/types.ts
@@ -1,0 +1,7 @@
+import { IconName } from '../Icon';
+
+export interface IMenuItem {
+  href: string;
+  i18nKey: string;
+  icon: IconName;
+}
